add primitive
