+++ primitives/core/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "selendra-core-primitives"
+version = "0.1.3"
+authors = ["Selendra <info@selendra.org>"]
+edition = "2021"
+
+[dependencies]
+scale-info = { version = "2.1.2", default-features = false, features = ["derive"] }
+parity-scale-codec = { version = "3.1.5", default-features = false, features = [ "derive" ] }
+parity-util-mem = { version = "0.11.0", default-features = false, optional = true }
+
+sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.26", default-features = false }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.26", default-features = false }
+sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.26", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+	"scale-info/std",
+	"parity-scale-codec/std",
+	"parity-util-mem",
+	"sp-core/std",
+	"sp-runtime/std",
+	"sp-std/std",
+]
+
